+++ Changelog.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+- Added support to bulk convert keys from a nip05 domain by [@snazzybytes]
+
 ## [0.2] - 2023-01-18
 ### Added
 - Docker image is now uploaded to GHCR by [@praveenperera]